requirements.txt
python-dotenv
openai>=1.40.0
numpy
pandas
tqdm
streamlit
